% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-generation.R
\name{Feature.Generation}
\alias{Feature.Generation}
\title{Feature Generation}
\usage{
Feature.Generation(V = 1e+06, LoB.dist = c(0.25, 0.3, 0.2, 0.25),
  inflation = c(0.01, 0.01, 0.01, 0.01), seed = 100)
}
\arguments{
\item{V}{Total expected number of claims.}

\item{LoB.dist}{Categorical distribution for the allocation of the claims to the four lines of business.}

\item{inflation}{Grow parameters (per LOB) for the numbers of claims in the 12 accident years.}

\item{seed}{Random seed.}
}
\description{
Feature Generation
}
