% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-machine.R
\name{Simulation.Machine}
\alias{Simulation.Machine}
\title{Simulate Claims}
\usage{
Simulation.Machine(features, npb = nrow(features), seed1 = 100,
  std1 = 0.85, std2 = 0.85)
}
\arguments{
\item{features}{Data frame with features for which we would like to model the cash flow patterns.}

\item{npb}{Number of observations that are treated at the same time (number per block).}

\item{seed1}{Seed for reproducibility.}

\item{std1}{Value of the standard deviation used in the log-normal distribution of the claim sizes.}

\item{std2}{Value of the standard deviation used in the log-normal distribution of the recovery sizes.}
}
\description{
Simulate Claims
}
\keyword{internal}
