% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjuror.R
\name{simulation_machine}
\alias{simulation_machine}
\title{Create a Simulation Machine Specification}
\usage{
simulation_machine(num_claims, lob_distribution, inflation, sd_claim,
  sd_recovery)
}
\arguments{
\item{num_claims}{Number of expected records to generate.}

\item{lob_distribution}{A length 4 vector. Distribution of claims among the
four LOBs.}

\item{inflation}{A length 4 vector. Grow parameters (per LOB) for the numbers
of claims in the 12 accident years.}

\item{sd_claim}{Value of the standard deviation used in the log-normal
distribution of the claim sizes.}

\item{sd_recovery}{Value of the standard deviation used in the log-normal
distribution of the recovery sizes.}
}
\value{
A \code{charm} object.
}
\description{
Create a Simulation Machine Specification
}
\examples{
charm <- simulation_machine(
  num_claims = 50000, 
  lob_distribution = c(0.25, 0.25, 0.30, 0.20), 
  inflation = c(0.01, 0.01, 0.01, 0.01), 
  sd_claim = 0.85, 
  sd_recovery = 0.85
)
charm

}
