% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjuror.R
\name{conjure.simulation_machine}
\alias{conjure.simulation_machine}
\title{Simulate Claim Histories}
\usage{
\method{conjure}{simulation_machine}(incantation, seed = NULL,
  rows_per_partition = NULL, ...)
}
\arguments{
\item{seed}{Seed for random number generation.}

\item{rows_per_partition}{(Optional) For parallel processing, number of 
observations that are treated at the same time.}
}
\description{
Simulate Claim Histories
}
