% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjuror.R, R/util.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{conjure}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{conjuror}{\code{\link[conjuror]{conjure}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

