% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjuror.R
\name{simulation_machine}
\alias{simulation_machine}
\title{Create a Simulation Machine Specification}
\usage{
simulation_machine(num_records, lob_distribution, inflation)
}
\arguments{
\item{num_records}{Number of expected records to generate.}

\item{lob_distribution}{A length 4 vector. Distribution of claims among the 
four LOBs.}

\item{inflation}{A length 4 vector. Grow parameters (per LOB) for the numbers
of claims in the 12 accident years.}
}
\value{
An incantation object.
}
\description{
Create a Simulation Machine Specification
}
